Move Track and metadata to model mod
