+++ src/rekordbox/mod.rs
@@ -38,6 +38,6 @@ pub use server::ApplicationEvent as Event;
 use rpc::server as rpc_server;
 use library::DBLibraryServer;
 pub use packets::DBMessage;
-pub use library::{Track, Metadata};
+pub use library::model::{Track, Metadata};
 pub use library::database::Database;
 use library::metadata_type;
